+++ box/Software/Triggering/src/ZMQSerializer.cpp
@@ -12,7 +12,7 @@ using std::pair;
 ZMQSerializer::ZMQSerializer() :  _ctx(),_client(_ctx, zmqpp::socket_type::pub) {
     Settings *set = Settings::Instance();
 
-    _boxId = set->getInt("box_id");
+    _boxId = set->getInt64("box_id");
     _idString = std::to_string(_boxId);
 
     std::string host = set->getString("zmq.trigger_host");;
